Adds start round command processor

+++ src/main/kotlin/com/artifactgames/copyplash/model/RequestModels.kt
@@ -15,6 +15,6 @@ data class GameMode(val mode: GameModes = GameModes.INSPIRATION, val rounds: Lis
     }
 }
 
-data class Round(val position: Int, val title: String, val description: String, val timeout: Int)
+data class Round(val position: Int, val title: String, val description: String, val timeout: Long)
 
 data class GameSettings(val locale:String)
